+++ .eslintrc.js
@@ -11,7 +11,7 @@ module.exports = {
   },
   parser: '@typescript-eslint/parser',
   parserOptions: {
-    project: ['./tsconfig.json'],
+    project: ['./tsconfig.eslint.json'],
     sourceType: 'module',
   },
   plugins: ['ban', 'filenames', 'local', '@typescript-eslint'],