+++ documentation/snippets/formula/boolean_formula.ts
@@ -1,0 +1,21 @@
+import * as coda from '../../../index';
+
+const pack = coda.newPack();
+
+// BEGIN
+
+pack.addFormula({
+  resultType: coda.ValueType.Boolean,
+  name: 'MyFormula',
+  description: '',
+  parameters: [
+    coda.makeParameter({
+      type: coda.ParameterType.String,
+      name: 'myParam',
+      description: '',
+    }),
+  ],
+  execute: async ([param]) => {
+    return param === 'true';
+  },
+});