+++ documentation/snippets/parameter/array/image_array_parameter.ts
@@ -1,0 +1,9 @@
+import * as coda from '../../../../index';
+
+// BEGIN
+
+coda.makeParameter({
+  type: coda.ParameterType.ImageArray,
+  name: 'myParam',
+  description: 'My description',
+});