+++ documentation/snippets/parameter/boolean_parameter.ts
@@ -1,0 +1,9 @@
+import * as coda from '../../../index';
+
+// BEGIN
+
+coda.makeParameter({
+  type: coda.ParameterType.Boolean,
+  name: 'myParam',
+  description: 'My description',
+});