+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.4.1
+
+The pack builder now includes a method `setDefaultConnectionRequirement()`
+which sets a default value (Required, Optional, or None) to apply
+to all formula definitions if an explicit value is not specific
+when those formulas are defined. This is very convenient when you
+are using authentication and you simply want every formula in your pack
+to use a connection (aka an account). This is just syntactic sugar
+that does the same thing as specifying a `connectionRequirement`
+on every formula and sync table definition. Example usage:
+
+```typescript
+export const pack = newPack();
+pack.setDefaultConnectionRequirement(ConnectionRequirement.Required);
+pack.addFormula(...);
+pack.addSyncTable(...);
+```
+
+## 0.4.0
+
+### TypedStandardFormula renamed to Formula (TypeScript only)
+
+The type `TypedStandardFormula`, which is the type used for the `formulas` array in main
+`PackVersionDefinition` type has been renamed `Formula` to be simpler and more intuitive.
+
 ## 0.3.1
 
 ### Metadata formulas no longer need to be wrapped in `makeMetadataFormula()`.